Pequeños cambios :D

+++ src/main/java/com/hospital_novasalud/hospital_nova_salud/controllers/DoctorController.java
@@ -3,8 +3,10 @@ package com.hospital_novasalud.hospital_nova_salud.controllers;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -17,6 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.hospital_novasalud.hospital_nova_salud.dto.DoctorDto;
 import com.hospital_novasalud.hospital_nova_salud.models.Doctor;
+import com.hospital_novasalud.hospital_nova_salud.models.Especialidad;
 import com.hospital_novasalud.hospital_nova_salud.services.IDoctorService;
 
 import jakarta.validation.Valid;
@@ -41,12 +44,24 @@ public class DoctorController {
         if(result.hasFieldErrors()) {
             return validation(result);
         }
-        return doctorService.save(doctor);
+        boolean existe = doctorService.save(doctor);
+        Especialidad especialidad = doctorService.findEspecialidad(doctor);
+
+        if(especialidad == null) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("No se encontró la especialidad");
+        if(!existe) return ResponseEntity.status(HttpStatus.CREATED).body("Se registro al doctor con exito");
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("El nombre de usuario para este doctor ya existe");
+
     }
 
     @DeleteMapping("/eliminar/{id}")
     public ResponseEntity<?> eliminarUsuario(@PathVariable Long id){
-        return doctorService.deleteById(id);
+        Optional<Doctor> doctor = doctorService.findDoctorById(id);
+        if(doctor.isEmpty() || doctor.get().getUsuario().getEstado().getId() == 2){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Doctor no encontrado");
+        }
+        
+        doctorService.deleteById(id);
+        return ResponseEntity.status(HttpStatus.OK).body("Doctor eliminado con exito");
     }
     private ResponseEntity<?> validation(BindingResult result) {
         Map<String, String> errors = new HashMap<>();
