+++ src/main/java/com/hospital_novasalud/hospital_nova_salud/controllers/UsuarioController.java
@@ -1,17 +1,14 @@
 package com.hospital_novasalud.hospital_nova_salud.controllers;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -19,7 +16,6 @@ import com.hospital_novasalud.hospital_nova_salud.dto.UsuarioDto;
 import com.hospital_novasalud.hospital_nova_salud.models.Usuario;
 import com.hospital_novasalud.hospital_nova_salud.services.IUsuarioService;
 
-import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("/api/usuarios")
@@ -31,24 +27,13 @@ public class UsuarioController {
     public List<UsuarioDto> listarUsuarios() {
         return usuarioService.findAll();
     }
-    @PostMapping("/registrar")
-    public ResponseEntity<?> registrarUsuarios(@Valid @RequestBody Usuario usuario, BindingResult result) {
-        if(result.hasFieldErrors()) {
-            return validation(result);
-        }
-        return usuarioService.save(usuario);
-    }
-
     @DeleteMapping("/eliminar/{id}")
     public ResponseEntity<?> eliminarUsuario(@PathVariable Long id){
-        return usuarioService.deleteById(id);
-    }
-
-    private ResponseEntity<?> validation(BindingResult result) {
-        Map<String, String> errors = new HashMap<>();
-        result.getFieldErrors().forEach(
-            err -> errors.put(err.getField(), "El campo " + err.getField() + " " + err.getDefaultMessage())
-        );
-        return ResponseEntity.badRequest().body(errors);
+        Optional<Usuario> usuario = usuarioService.findUsuarioById(id);
+        if(usuario.isEmpty() || usuario.get().getEstado().getId() == 2){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuario no encontrado");
+        }
+        usuarioService.deleteById(id);
+        return ResponseEntity.status(HttpStatus.OK).body("Usuario eliminado con exito");
     }
 }
